+++ jsonargparse/core.py
@@ -14,7 +14,7 @@ from unittest.mock import patch
 from .formatters import DefaultHelpFormatter, empty_help, formatter_context, get_env_var
 from .jsonnet import ActionJsonnet
 from .jsonschema import ActionJsonSchema
-from .loaders_dumpers import check_valid_dump_format, dump_using_format, get_loader_exceptions, loaders, load_value, load_value_context
+from .loaders_dumpers import check_valid_dump_format, dump_using_format, get_loader_exceptions, loaders, load_value, load_value_context, yaml_load
 from .namespace import is_meta_key, Namespace, split_key, split_key_leaf, strip_meta
 from .signatures import is_pure_dataclass, SignatureArguments
 from .typehints import ActionTypeHint, LazyInitBaseClass
@@ -161,7 +161,7 @@ class ArgumentParser(_ActionsContainer, argparse.ArgumentParser):
         print_config: Optional[str] = '--print_config',
         parser_mode: str = 'yaml',
         default_config_files: Optional[List[str]] = None,
-        default_env: bool = False,
+        default_env: bool = yaml_load(os.environ.get('JSONARGPARSE_DEFAULT_ENV', 'False')),
         default_meta: bool = True,
         **kwargs
     ):